+++ src/components/CardCredits.jsx
@@ -1,6 +1,5 @@
-import React, {useContext, useEffect} from 'react';
-import {useParams} from 'react-router-dom';
-import MaratoneiContext from '../context/MaratoneiContext';
+import React from 'react';
+import PropTypes from 'prop-types';
 import {v4 as uuidv4} from 'uuid';
 import {
   CardCreditsStyled,
@@ -8,19 +7,14 @@ import {
   PersonStyled,
 } from '../styles/index';
 
-function CardCredits() {
-  const {castAndCrewData, fetchCastAndCrew} = useContext(MaratoneiContext);
-  const {id} = useParams();
-
-  useEffect(() => {
-    fetchCastAndCrew(id);
-  }, []);
+function CardCredits({detailsData}) {
+  const {cast_and_crew: castAndCrew} = detailsData;
 
   return (
     <CardCreditsStyled>
       <h2>Quem ta participando?</h2>
       <CastAndCrew>
-        {castAndCrewData.map((person) => {
+        {castAndCrew.map((person) => {
           const {
             name,
             character,
@@ -44,5 +38,9 @@ function CardCredits() {
     </CardCreditsStyled>
   );
 }
+
+CardCredits.propTypes = {
+  detailsData: PropTypes.object,
+};
 
 export default CardCredits;
