+++ src/context/MaratoneiProvider.js
@@ -1,37 +1,18 @@
 import React, {useState} from 'react';
 import PropTypes from 'prop-types';
 import MaratoneiContext from './MaratoneiContext';
-import castAndCrew from '../services/castAndCrewAPI';
-import movieReviews from '../services/movieReviewsAPI';
-import useRequestAPI,
-{
-  useQueryAPI,
-} from '../hooks/useRequestAPI';
+import useRequestAPI, {useQueryAPI} from '../hooks/useRequestAPI';
 
 function MaratoneiProvider({children}) {
   const {result, request, setRequest, setResult} = useRequestAPI();
   const setFilterQuery = useQueryAPI(request, setResult);
 
-  const [castAndCrewData, setCastAndCrewData] = useState([]);
-  const [reviews, setReviews] = useState([]);
   const [toggleFilter, setToggleFilter] = useState('stand-by-toggle');
   const [rotateWhenClick, setRotate] = useState('stand-by');
   const [filter, setFilter] = useState({
     inputSearchFilter: '',
     yearSearchFilter: '',
   });
-
-  const fetchReviews = async (movieId) => {
-    const data = await movieReviews(movieId);
-
-    return setReviews(data);
-  };
-
-  const fetchCastAndCrew = async (movieId) => {
-    const data = await castAndCrew(movieId);
-
-    setCastAndCrewData(data);
-  };
 
   const handleToggle = () => {
     if (rotateWhenClick === 'stand-by') setRotate('filterActivated');
@@ -83,10 +64,7 @@ function MaratoneiProvider({children}) {
     <MaratoneiContext.Provider value={ {
       result,
       setRequest,
-      fetchCastAndCrew,
-      fetchReviews,
       handleToggle,
-      reviews,
       toggleFilter,
       rotateWhenClick,
       filter,
@@ -94,7 +72,6 @@ function MaratoneiProvider({children}) {
       handleClickSearch,
       handleEnterSearch,
       removeFilters,
-      castAndCrewData,
     } }>
       {children}
     </MaratoneiContext.Provider>
