+++ src/services/requestDetails.js
@@ -17,12 +17,26 @@ async function fetchMovieDetail(showId) {
         .then((response) => response.json())
         .then((response) => response.results.BR);
 
+    const requestReviews = await fetch(`https://api.themoviedb.org/3/movie/${showId}/reviews?api_key=${TMDB_API_KEY}&language=pt-BR&page=1`)
+        .then((response) => response.json())
+        .then((response) => response.results);
+
+    const requestCastAndCrew = await fetch(`https://api.themoviedb.org/3/movie/${showId}/credits?api_key=${TMDB_API_KEY}&language=pt-BR`)
+        .then((response) => response.json());
+
+    const actors = requestCastAndCrew.cast.slice(0, 6);
+    const castAndCrew = [requestCastAndCrew.crew
+        .find((crew) => crew.department === 'Directing')]
+        .concat(actors);
+
     requestMovieDetails['trailer_key'] = requestTrailerVideo;
     requestMovieDetails['recomendations'] = requestRecomendations.slice(0, 4);
     requestMovieDetails['providers'] = requestProviders;
+    requestMovieDetails['reviews'] = requestReviews;
+    requestMovieDetails['cast_and_crew'] = castAndCrew;
     requestMovieDetails.production_companies = requestMovieDetails
         .production_companies.filter((comp) => comp.logo_path !== null);
-
+    console.log(requestMovieDetails);
     return requestMovieDetails;
   } catch (error) {
     console.log(error);
